☕️ 优化事件回调处理
