+++ ios/RNAgoraExample.xcodeproj/xcshareddata/xcschemes/RNAgoraExample.xcscheme
@@ -99,6 +99,13 @@
             ReferencedContainer = "container:RNAgoraExample.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
+      <EnvironmentVariables>
+         <EnvironmentVariable
+            key = "OS_ACTIVITY_MODE"
+            value = "disable"
+            isEnabled = "YES">
+         </EnvironmentVariable>
+      </EnvironmentVariables>
       <AdditionalOptions>
       </AdditionalOptions>
    </LaunchAction>
